aggiustata la produzione del seed

+++ PMCSN_nuovo/src/main/java/org/uniroma2/PMCSN/utils/IntervalCSVGenerator.java
@@ -109,7 +109,7 @@ public class IntervalCSVGenerator {
     }
 
     public static void writeGlobalInterval(
-            int rep,
+            long rep,
             double reportTime,
             List<SimpleMultiServerNode> localNodes,
             String baseDir
